On the 1.7.x-issue4153 branch, fix summarize diff for added and deleted diff targets. This is a little more involved than I had hoped.
Partially merge r1164116 from trunk. That revision renamed the
file libsvn_client/repos_diff_summarize.c to libsvn_client/diff_summarize.c.
The merge added libsvn_client/diff_summarize.c as a new file on this branch,
and produced an "local edit vs. incoming delete" tree-conflict on
repos_diff_summarize.c which was trivially resolved by reverting this file.
The intention is to keep both files alive on this branch.

The new diff_summarize.c file contains summarize diff callbacks which
can be driven by the code that handles added and deleted diff targets.

To make these callbacks accessible within libsvn_client, declare the
function svn_client__get_diff_summarize_callbacks() in libsvn_client/client.h.

On trunk, svn_client__get_diff_summarize_callbacks() replaced the
function svn_client__get_diff_summarize_editor(), but in 1.7 we want
to keep the latter function around to minimize changes made to the
summarize diff code.

Also, effectively merge r1338739 from trunk in this commit to fix
instantiation of the summarize diff callbacks.

* subversion/libsvn_client/diff.c
  (diff_summarize_repos_repos): Handle added and deleted diff targets
   by calling diff_repos_repos_added_or_deleted_target() if one of
   the diff targets does not exist.

* subversion/libsvn_client/client.h
  (svn_client__get_diff_summarize_callbacks): Declare.


git-svn-id: https://svn.apache.org/repos/asf/subversion/branches/1.7.x-issue4153@1338753 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_client/diff.c
@@ -2488,6 +2488,27 @@ diff_summarize_repos_repos(svn_client_diff_summarize_func_t summarize_func,
                                    revision1, revision2,
                                    peg_revision, pool));
 
+  if (kind1 == svn_node_none || kind2 == svn_node_none)
+    {
+      svn_wc_diff_callbacks4_t *callbacks;
+      void *callback_baton;
+
+      /* One side of the diff does not exist.
+       * Walk the tree that does exist, showing a series of additions
+       * or deletions. */
+      SVN_ERR(svn_client__get_diff_summarize_callbacks(
+                &callbacks, &callback_baton, target1,
+                summarize_func, summarize_baton, pool));
+      SVN_ERR(diff_repos_repos_added_or_deleted_target(target1, target2,
+                                                       rev1, rev2,
+                                                       kind1, kind2,
+                                                       callbacks,
+                                                       callback_baton,
+                                                       ra_session,
+                                                       pool));
+      return SVN_NO_ERROR;
+    }
+
   /* Now, we open an extra RA session to the correct anchor
      location for URL1.  This is used to get the kind of deleted paths.  */
   SVN_ERR(svn_client__open_ra_session_internal(&extra_ra_session, NULL,
